c++ / mojo changes for mus_demo in 'external window mode'
This CL is the first step towards having 'external window mode' support
in Mus. The existing MusDemo is adapted to test the functionality,
and its respective unit test (mus_demo_unittests) is changed to
exercise  multiple external windows creation in a follow up CL [2].

[2] https://codereview.chromium.org/2715533005/

New code flow:

Note: In Chrome, it also happens in
c/b/ui/views/chrome_browser_main_extra_parts_views.cc,
::ServiceManagerConnectionStarted, when MusClient is created.

* MusDemoExternal calls WindowTreeClient::CreatesForExternalWindowMode.
This creates and holds a WindowTreeClient instance in 'external window mode'.
Additionally, it also creates a ExternalWindowModeRegistrar instance.
Yet from ::CreateForExternalWindowMode ExternalWindowModeRegistrar::Register
is called, creating WindowTree and an ExternalWindowTreeHostFactory instances
on the server side, and acquiring mojo handles on the client side.

Note: In external window mode, the WindowTree and ExternalWindowTreeHostFactory
are unique instances, serving 0 or 'n' WindowTreeHostMus instances.

* Yet from MusDemoExternal, WindowTreeHostMus instance(s) are
created, in accordance to the value passed to command line
parameter --external-window-count, or 1 by default.

Note: In Chrome, this happens in c/b/ui/views/frame/browser_frame_mus.cc,
::GetWidgetParams when DesktopWindowTreeHostMus is created for Chrome
outer/platform window.

WindowTreeHostMus ctor takes a WindowTreeClient instance which
works as a "window tree host delegate".

* During the creation chain of WindowTreeHostMus, WindowTreeClient::OnWindowMusCreated
is called, and this is what calls the newly added method
ExternalWindowTreeHostFactory::CreatePlatformWindow, when in 'external
window mode'. That is the entry point of of ws::Display instance
creation.

* In the existing creation flow of ws::Display, the server
actually calls back to the client via WindowTreeClient::OnEmbed
when it is done. In 'external window mode' this callback is kept,
but it passes a 'null' WindowTree object as parameter.
Why? Because in ::CreateForExternalWindowMode (above) the
unique WindowTree instance was already created.

* services/ui/ws/external_window_mode_registrar.cc|h:

  This class implements ExternalWindowModeRegistrar class.
  It exposes a ::Register method, that:

    (a) Binds the mojom::WindowTreeHostFactoryRequest
    (b) Binds the mojom::WindowTreeRequest
    (c) Stores the mojom::WindowTreeClientPtr to call out
        WindowTreeClient later on.
    (d) A ws::WindowTree instance is also created here. It ensures
        single WindowTree and WindowTreeClient instances serving
        various WindowTreeHost instances.

  The goal of the ExternalWindowModeRegistrar interface is
  to ensure a WindowTreeHostFactory::CreatePlatformWindow can not
  be called before WindowTreeHostFactory being properly set up.

* services/ui/ws/window_tree_host_factory.cc|h:

  Adds ::CreatePlatformWindow method so that it triggers the
  creation of per-outer-window ws::Display instances in external
  window mode.

TODO (in follow up CLs):

  1) Use ExternalWindowTreeHostFactory as an associated interface
    from WindowTreeClient. An instance would be acquired via
    WindowTree::GetWindowTreeHostFactory.
    This allows both interfaces to share the same mojo pipe, ensuring
    ordering of messages. This is important because during the creation
    of windows/ws::Display's, ExternalWindowTreeHostFactory::CreatePlatformWindow
    messages concurs with subsequent WindowTree method calls, causing
    non-deterministic crashes in chrome/mus.
  2) Rename WindowManagerDisplayRoot since it is used in non-WindowManager
  path.
  3) Clean up mus_demo_external (launch instances in parallel).
  4) Clean up WindowTreeClient::OnEmbed flow.
  5) Launching Chrome/Mus in external window mode, on Linux.

BUG=666958