fixup! Handle OnBoundsChange in external mode. (#8)
Fix menu windows placement, when root window (main chrome window) is
moved or resized.

More or less, this commit is mostly about proper handling of changed
bounds sent from window manager. Additional methods were implemented in
order to make it possible to send bounds to StubWindow, which is used by
other windows as a reference for their bounds. But, the main point here
is that server_window, which is root_ in ui::ws::Display, must not
change its origin as long as its origin is relative to the
X11WindowOzone rather than to a real screen.