Allow ws to pass display::Display data to Aura/client
Change ScreenManagerOzoneExternal to be able to send out display::Display
updates via ws::UserDisplayManager (which will send the display::Display
to the client).

For this, patch extends OzonePlatform with a method that allows
its instances (X11, Wayland, etc) to query the host system its displays
data (resolution only for now).
ScreenManagerOzoneExternal then calls to ws::DisplayManager, which calls to
ws::UserDisplayManager.

In order to allow ws::DisplayManager to get display::Display and not change
ws::Display in external mode, a new hook was added: ::OnHostDisplaysReady.

Wayland support will come in a fixup.

Issue #43