fixup! Allow ws to pass display::Display data to Aura/client
PR #50, reverted an initial hack added as part of the commit
"Allow chrome launch with --mus parameter", but it left over one
bit.

This commits removes the left over bit, but in the next rebase
this should be squashed against "Allow chrome launch with --mus parameter"
commit, not "Allow ws to pass display::Display data to Aura/client".