fixup! c++ / mojo changes for 'external window mode'
In external window mode, it makes no sense to call
DisplayManager::OnDisplayUpdate when a ws::Display is
initialized.

First because ws::Display has not information about
display::Display instances. Second because ws::Display::display_
is solely to be used in internal window mode, as per the comment
in the header.