[ozone/wayland] Fix Showing/Hiding, Placing and Events for Menu Windows (#73)
This adaptation makes it possible to show/hide and place menu windows
properly in regards to their parent surface. What is more, this patch
fixes event locations sent to the client from wayland.
Conflicts:
	ui/ozone/platform/wayland/wayland_window.cc

fixup! [Wayland] Fix Showing/Hiding, Placing and Events for Menu Windows (#73)

This resolves crash when a user requests to open a submenu window

fixup! [Wayland] Fix Showing/Hiding, Placing and Events for Menu Windows (#73)

Typo fix

fixup! [Wayland] Fix Showing/Hiding, Placing and Events for Menu Windows (#73)

fixup! [ozone/wayland] Fix Showing/Hiding, Placing and Events for Menu Windows (#73)

Fix crash when menus are destroyed, but aura is still trying to hide them.
It may happen when wayland decides to destroy the menu (wrong serial number, for example).
The problem here is that we set a child window pointer to the parent, but never remove it.

fixup! [ozone/wayland] Fix Showing/Hiding, Placing and Events for Menu Windows

Ensure that a pointer to child window is invalidated as soon as xdg_popup_ is
destroyed. This fixes a random crash, when the wrong pointer to a child
deletes not the top most popup window, which results in a dismiss signal from
wayland and all windows are being closed.