fixup! First attemp to remove decorations from small windows
This method extends the existing ws::DisplayViewport struct to
hold information about the window type as well.
This allows the underlaying window manager to create the proper
host window type.

Implementation helps both the ozone/x11 and ozone/wayland backends.
The former uses a hardcoded value to hide (or not) window manager
window decorations (see the omnibox widget - URL autocompletion).
The later, creates "popups" based on a hardcoded window size.

Note that the method added PlatformWindowDelegate has an out parameter
and a default implementation. This is to avoid an empty method body in
the header, and not to have to add implementations to all classes that
inherit from this PlatformWindowDelegate.

Issue #38