Handle window minimize/maximize/restore in external window mode.
This patch makes it possible to change states of windows by an event sent
from a client or from a server.

The call flow is the following - as soon as WindowTreeClient::SetProperty
is called, new properties are passed to the ServerWindow, which notifies
WindowServer about the changes. The WindowServer checks type of the property
and calls PlatformDisplayDefault::SetNativeWindowState, which
decides whether to minimize/maximize or restore the window.

As soon as a native window server receives a request to change a state of
the request window, it processes the call and send a PropertyChanged event,
which is processed and sent upwards to WindowTreeClient::OnWindowStateChanged,
which processes the call and calls WindowTreeHostMus and aura::Window to
change the property if it has not been changed yet and show or hide the window.

Issue #75

fixup! Handle window minimize/maximize/restore in external window mode.

Fix compilation failures

fixup! Handle window minimize/maximize/restore in external window mode.

Add support for fullscreen mode in X11

fixup! Handle window minimize/maximize/restore in external window mode

Fix Wayland minimize/maximize/restore/fullscreen modes.
It's still bogus and doesn't work correctly as long as wayland
makes windows to be misplaced after unmaximize or unfullscreen.
What is more, this patch adds support to handle events from client only.
Events like restoring a window from a tray are not supported yet.