fixup! Handle OnBoundsChange in external mode.
This patch changes a bit the way we set bounds from
the server.

Previously, we were not setting the WindowServer bounds
directly, and instead the server was calling the client via
WindowTreeClient::OnNewBoundsFromHostServer, which called
back to the server with WindowTreeHostMus::SetBoundsInPixels
and WindowTreeClient::OnWindowTreeHostBoundsWillChange.
At this point, we were resetting the X,Y values to 0,0
so that mouse events did not break.
When WindowTree::SetWindowBounds were called, the WindowServer
bounds were set accorgindly.

Now, when bounds are set from the server, we set the WindowServer
bounds directly, making sure that:

- Display::root_ has the actually bounds set from the server.
- WindowManagerDisplayRoot::root_ has rect(0,0, size) set, not
to break mouse events.

Additionally, we call out the client via WindowTreeHostMus::SetBoundsFromServer,
so that it prevents this calling back into the server.

Issue #86