Handle events for interactive drag and resize of native windows.
This patch adds capabilities to process events, which are aimed
at starting of interactive drag and resize of native windows.
Class X11WindowEventFilter is now inherits from WindowEventFilter,
which is a main class for both Ozone X11/Wayland and non-ozone X11 builds.
That class handles the above mentioned events and can send maximize/restore/
resize/drag events down to window manager in ozone builds. In case of
non-ozone X11 builds, the move or resize events are handled by
X11WindowEventFilter and WM_MOVERESIZE atom is set (as it originally was).

At the moment, ozone X11 build doesn't have the final
PerformNativeWindowDragOrResize() method implemented as long as
the location of events is faulty there a bit

Conflicts:
	ui/views/widget/desktop_aura/window_event_filter.cc
	ui/views/widget/desktop_aura/window_event_filter.h
	ui/views/widget/desktop_aura/x11_window_event_filter.cc