fixup! c++ / mojo changes for 'external window mode'
This CL makes the WindowTreeClient awaits on the WindowTree
connection to establish before it continues.

Before this, there was a mismatch between the values of
WindowTreeClient::client_id_ and the WindowTree::id_. The
reason was because the former was only updated when
WindowTreeClient::OnEmbed was called.

With the patch, we behave similarly to (m)ash, where it awaits for
the WindowTreeClient::OnConnect call, where the WTC::client_id_ is
set.

This fixes one of the blockers on using WindowManagerAccessPolicy
for both chrome/mus_demo and unittests.