fixup! Handle window minimize/maximize/restore in external window mode.
Fix ozone_unittests and make some tweaks in WaylandWindow:

Tests are now modified in such a way that the window's state is
restored only after configure event is sent and delegate is said about
the window's state changed. This emulates real client->wayland->client
behaviour much better.

What is more, WaylandWindow now tells about FULLSCREEN state as well,
but delegate checks that and just doesn't notify mus about that, because
it was mus, who instantiated that. This is made on purpose to make
the job to write tests easier. Otherwise, MessageLoops and other tweaks
are needed.