fixup! c++ / mojo changes for 'external window mode'
Another preparation work for the refactor patch

Basically, it calls WindowTree::AddRoot from WindowTreeHostFactory
rather than from Display::InitRootDisplay. This brings in no functionality
change, but ensures we have a DefaultPlatformDisplay instance fully
instantiated when ::AddRoot is called, something that is important in
case we want to adopt the refactor branch.