Rework server WindowTree hierarchy, so match ChromeOS'
This CL reworks the WindowTree hierarchy of our external
window mode support, to make ChromeOS's mash/mus.

Essentially, the tree used to look like:

  Display::root_
    WindowManagerDisplayRoot::root_ which hosts a BrowserFrame aura window)

whereas it now it now looks like

  Display::root_
    WindowManagerDisplayRoot::root_
      Root ServerWindow (which hosts a BrowserFrame aura window)

CL also adapts the server window id's definition in order to keep
tests failing. Today, WindowTreeClientTests assume no root server
window underneath WindowManagerDisplayRoot::root. In order to overcome
this, we are creating the server window root, with a specific Id, when
running tests.

Ideally, this commit should be split up in various fixups, once we
define this is the design we want to follow, but for now leave it
separate.