[ozone/wayland] Add xdg_v6 support
This commit adds support for xdg_v6 by having a wrapper for different
xdg shell versions.

The interfaces xdg_popup_wrapper and xdg_surface_wrapper are
implemented by different xdg objects with different versions.
WaylandWindow doesn't know what version is used and relies on
XDGShellObjectsFactory, which decides which version of objects
to use by checking the shell version. What is more, this design
should allow to add ivi shell support quite easily in the future.

The decision of what xdg shell version to connect to is the following:
when WaylandConnection::Global is called back,
it checks what shell versions are available. The first version to
be checked is V6. If V6 is not available, V5 is checked and
the interface is binded. Otherwise V6 is used.

fixup! [ozone/wayland] Add xdg_v6 support

Move creation of xdg surface and popups to own functions