[chrome/mus/linux] Restore previous session's bounds and state of a window
This patch makes it possible to store and restore bounds and window state
so that chrome/mus behaves exactly the same way as stock X11.

In details: two problems are fixed with this patch.
1) BrowserView::SaveWindowPlacement() had always been failing before,
because frame_->ShouldSaveWindowPlacement() (or BrowserFrameMus)
had false value hardcoded to be returned. Now we have a check, which ends up with
true if kRestoreBoundsKey is not stored in the aura::window's properties.
2) BrowserFrameMus::GetWindowPlacement(), which is used by Widget::SaveWindowPlacement, where
native_widget_->GetWindowPlacement() is called, had always been returning
hardcoded bounds (10, 10, 800, 600). Now bounds are taken from |content_window_|
and kRestoreBoundsKey is stored, restored bounds and restored state are taken from gfx::NativeWindow.

Issue: #162