Support starting chromium with URL on Linux/Ozone build
This patch makes it possible to run chrome the following way:
./chromium [URL]

Call to AdjustServiceProcessCommandLine() is intented to remove
all the not necessary switches, which could lead a process
to become a ServiceManager. But it also strips out all the args,
because those are without switches. This is done
unintentionally. Thus, all the args without switches are
stored if (identity.name() == content::mojom::kPackagedServicesServiceName)
resolves to true and then args are appended back.