fixup! c++ / mojo changes for 'external window mode'
Avoid double initialization of WindowTreeHost: after
https://chromium-review.googlesource.com/c/606933 commit, the
chromium browser on Linux build with ozone has stopped to
start. It has always been observing a crash due to a DCHECK
in WindowTreeHost::InitCompositor(), which checked that
the compositor must not have a root_layer() set.

The reason of hitting the DCHECK is that the very first initialization
is always done during the browser start up:

6 0x7f4ff06c0981 views::DesktopNativeWidgetAura::InitNativeWidget()
7 0x7f4ff0682405 views::Widget::Init()
8 0x55e62a03742c BrowserFrame::InitBrowserFrame()
9 0x55e629d227e9 BrowserWindow::CreateBrowserWindow()
10 0x55e629b3d4e5 (anonymous namespace)::CreateBrowserWindow()
11 0x55e629b3cb79 Browser::Browser()
12 0x55e629b9b1b2 StartupBrowserCreatorImpl::OpenTabsInBrowser()
13 0x55e629b9cffa StartupBrowserCreatorImpl::RestoreOrCreateBrowser()
14 0x55e629b9a6a6 StartupBrowserCreatorImpl::ProcessLaunchUrlsUsingConsolidatedFlow()
15 0x55e629b99742 StartupBrowserCreatorImpl::Launch()
16 0x55e629b93a72 StartupBrowserCreator::LaunchBrowser()
17 0x55e629b929d4 StartupBrowserCreator::ProcessCmdLineImpl()
18 0x55e629b91102 StartupBrowserCreator::Start()
19 0x55e626f60fc7 ChromeBrowserMainParts::PreMainMessageLoopRunImpl()
20 0x55e626f5f8bc ChromeBrowserMainParts::PreMainMessageLoopRun()
21 0x7f4ff563c574 content::BrowserMainLoop::PreMainMessageLoopRun()

After this when WindowTreeClient::OnEmbed() is called,
we used to call window_tree_host->InitHost() second time and then
crash.

What is more, in order to mimic standart browser creation,
make WindowTreeDataExternal to call InitHost() before allowing
to calling to create an actual window.