fixup! [ozone/wayland] Add xdg_v6 support
Wayland can ask to a client to set window geometry in certain cases
like before interactive drag/resize, unmaximize and etc.
If width and height are send as 0s, it means that a client
must answer Wayland with SetWindowGeometry command and acknowledge
that. If it is not done, certain compositors like Gnome with Wayland
can treat this not answered request as the client wants the window
to be with 0 widht and 0 height, which could result in a tiny window
during initial resize event.