Fix maximize/restore glitch of OpaqueBrowserFrameView in Weston.
This patch avoids spamming aura with states unless previous state
has been "minimized" state and makes the wayland client to decide
what bounds to use and send these back to aura, which
calls relayout and updates OpaqueBrowserFrameView.

fixup! Fix maximize/restore glitch of OpaqueBrowserFrameView in Weston.

Make the opaque view layout to be updated once window states are
updated upon the host window manager events like if a user moves a window
in such a way so that it becomes automatically maximized or restored.

The layout update behaviour is copied from DesktopWindowTreeHostX11::OnWmStateChanged.
Now DesktopWindowTreeHostMus subscribes for a root window, which receives
window state updates, in addition to the content window, and invalidates
and updates client and non-client views once aura::client::kShowStateKey
key is received.

What is more, in order to continue to avoid flooding aura, |restored_bounds_in_pixels_|
or |previous_bounds_in_pixels_| are tracked. If those are empty, then
it must be the host window manager, which triggered the change. Otherwise,
it has been the client and there is no need to flood it with states it already
knows.

Issue #270