First attempt to reduce Gpu service start up failure.
It happens occasionally that the gpu service (GpuServiceImpl) fails to
notify the gpu host (GpuHost) that it has been initialized
(in practice, GpuServiceImpl::InitializeWithHost calls out to
GpuHost::DidInitialize, via mojo, but it never reaches the end).

The natural flow when the call succeeds is:

- GpuHost notifies the WindowServer that Gpu services is initialized,
and then ui::Service.
- ui::Service would initialized ScreenManager (ScreenManagerOzoneExternal::Init),
which would call out to Ozone to query "host display info".
- Ozone would then call back to ScreenManagerOzoneExternal, then DisplayManager
and UserDisplayManager.
- The last would call out to ScreenMus, which awaiting in a sync call.

However, all this does not happen because the GpuService when does not
initialize well at the beginning, ending up in a dead wait.

The CL tries to avoid the sync wait at ScreenMus side, installing a
default display::Display, that would be replaced when Ozone sends the
proper display down the road.

Issue #212