fixup! Allow ws to pass display::Display data to Aura/client
Fix unittests: after the change to WaylandConnection::Initialize, the
unittests started to hang on startup, because the WaylandOutput::Geometry
had always been empty.

The fix is to create a MockOutput, which is a global wl object on the
server side, with non-empty geometry, which will be set during initializtion
step.

The second fix is to modify WaylandConnectionTest::Output in such a way that
it would not wait in a blocking run loop. At the moment, we have
a wl_display_roundtrip run until WaylandOutput::Geometry is set instead.