fixup! c++ / mojo changes for 'external window mode'
Factor out bits of WindowTreeClient::OnTopLevelCreated
into a helper ::OnTopLevelCreatedImpl method and call it
from ::OnEmbed (instead of ::ConfigureWindowDataFromServer).

In fact, ::OnTopLevelCreated is a more appropriate callback
to be called from the server, when a platform window is created
in external window mode. Today we call ::OnEmbed for convenience,
since unittests rely on it.

Reason: ::OnTopLevelCreated already handles the case of window
bounds or visibility changes between WindowTreeHostFactory::CreatePlatformWindow
and the associated WindowTreeClient callback is called.
::ConfigureWindowDataFromServer does not, possibly overriding
client's data wrongly.

This is specially needed for when WindowTreeHostFactory::CreatePlatformWindow
is called without the 'initial bounds'. In this case, we assign a
temporary bounds at creation, which is reported to the client, via ::OnEmbed.
If by the time OnEmbed is called, there is already a ::SetWindowBounds
call to the server, the client already has valid bounds.

This helps with tooltips and "task manager" windows.

Issue #86
Issue #260